requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
biopython